arm64: fpsimd: Use wxN aliases for fpsimd_{save,restore}
For historical reasons, the 'fpsimd_save' and 'fpsimd_restore' use an
unusual convention where the temporary register argument is provided as
an index 'N', rather than a register name 'xN' or 'wN'. This is painful
to read, and where the macros are used, it's not obvious that the
numeric argument is actually a register in disguise.

We introduced a solution to this back in commit:

  4c4dcd3541f8 ("arm64: assembler: introduce wxN aliases for wN registers")

That conmmit added a 'wxN' alias for each 'wN' register, allowing for
easy conversion from 'xN' to 'wN'.

Use these in 'fpsimd_save' and 'fpsimd_restore', and update callers
accordingly. In fpsimd_save_state() and fpsimd_load_state(), I have
changed the temporary register from x8 to x1 for consistency with other
fpsimd code, but this should have no functional effect as these assembly
functions only take a single parameter in x0.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>