WIP: Factor out MIDR definitions
Factor the MIDR.{Implementer,PartNum} values out into a separate
midr-defs.h header so that these can be used independently of any
macros (and can be safely shared by kernel code and tools code that
lives in the kernel tree).

This deliberately only contains the Implementer and PartNum values, and
deliberately avoids the MIDR_FUJITSU_ERRATUM_* definitions which are
only needed for kernel-side errata workarounds.

In theory we could generate the midr-defs.h similarly to sysreg-defs.h,
e.g. from something like:

	Implementer	0x41	ARM
		Part	0xD00	FOUNDATION
		Part	0xD03	CORTEX_A53
		Part	0xD04	CORTEX_A35
		Part	0xD05	CORTEX_A55
	EndImplementer

	Implementer	0x42	BRCM
		Part	0x100	BRAHMA_B53
		Part	0x516	VULCAN
	EndImplementer

... generating:

	#define ARM_CPU_IMP_ARM			0x41
	#define ARM_CPU_PART_FOUNDATION		0xD00
	#define ARM_CPU_PART_CORTEX_A53		0xD03
	#define ARM_CPU_PART_CORTEX_A35		0xD04
	#define ARM_CPU_PART_CORTEX_A55		0xD05

	#define ARM_CPU_IMP_BRCM		0x42
	#define BRCM_CPU_PART_BRAHMA_B53	0x100
	#define BRCM_CPU_PART_VULCAN		0x516

With that, the scripting could enforce ordering by the
Implementer/PartNum values (and consistent use of hex), which would
avoid accidental duplicates and make merge conflicts easier to deal
with.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>