kselftest/arm64: Don't corrupt P0 in SSVE-VL-* irritator handler
Where SVE is not present or SVE is disabled, the fp-stress SSVE-VL-*
tests unexpectedly fail with a SIGILL. This happens because the
irritator handler tries to execute a 'ptrue p0.d' instruction in
non-streaming mode, which is not generally sound:

* When SVE is not implemented, the instruction is UNDEFINED in
  non-streaming mode, and the instruction will trigger an "undef"
  exception (ESR_ELx.EC==0b000000).

* When SVE is implemented, but the kernel does not use SVE (e.g. when
  disabled on the kernel command line via "arm64.nosve"), the kernel
  will configure CPACR/CPTR to trap non-streaming SVE, and the
  instruction will trigger an SVE trap (ESR_ELx.EC==0b011001).

In either of these cases, the kernel will handle the resulting exception
by sending a SIGILL signal to the task.

The irritator began to execute the non-streaming 'ptrue p0.d' in commit:

  3e360ef0c0a1 ("kselftest/arm64: Corrupt P0 in the irritator when testing SSVE")

... under the mistaken assumption that this was benign.

Fix this by restoring the original logic, with 'rdffr p0.b' under
ifdeffery.

This reverts commit 3e360ef0c0a1fb6ce9a302e40b8057c41ba8a9d2.

Fixes: 3e360ef0c0a1 ("kselftest/arm64: Corrupt P0 in the irritator when testing SSVE")
Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Mark Brown <broonie@kernel.org>
Cc: Shuah Khan <shuah@kernel.org>
Cc: Will Deacon <will@kernel.org>
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-kselftest@vger.kernel.org
Cc: stable@vger.kernel.org