kselftest/arm64: Fix fp-stress sub-test exit status
The barf functions in the fp-stress sub-tests erroneously place the
status value into x1 (rather than x0) when invoking the exit syscall,
potentially returning a different status than expected.

By coincidence, the low 8 bits of x0 happen to be non-zero for all of
these cases, and the exit will be treated as a failure by the fp-stress
harness, which is the desired behaviour. Hence this doesn't lead to
failing tests being reported as passing.

In all cases x0 contains either:

* The result of the prior write syscall made by the puts macro, which is
  a write of the two character string "]\n".

* The result of the prior write syscall made by the putdecn() function,
  which is a write of the single-character '\n'.

In either case, the return value of that write syscall may be a non-zero
number of bytes written, or a negative error code. As the relevant
strings are fewer than 256 characters, and all (positive) error codes
are less than 256, the value in the low 8 bits of x0 will not be zero.

Fix the syscall invocations to pass the status value in x0. For
consistency with other syscall invocations in the tests, the argument
registers are loaded into registers in order, before the syscall number
is loaded into x8.

Fixes: 5e992c638ea5 ("selftests: arm64: Add stress tests for FPSMID and SVE context switching")
Fixes: 5aa45cc5355d ("kselftest/arm64: Add stress test for SME ZA context switching")
Fixes: 1c07425e902c ("kselftest/arm64: Add a stress test program for ZT0")
Fixes: 11a7a42ea76e ("kselftest/arm64: Validate SVCR in streaming SVE stress test")
Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Mark Brown <broonie@kernel.org>
Cc: Shuah Khan <shuah@kernel.org>
Cc: Will Deacon <will@kernel.org>
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-kselftest@vger.kernel.org