kselftest/arm64: Remove dead code
Several of the fp-stress sub-tests have commented-out code for a
"fpsimd.c acitivty log dump hack" (sic), which are apparently leftovers
from development.

The presence of the commented-out code makes the code harder to read,
and doesn't serve any useful purpose. Remove it for now.

This is a removal of dead code. There should be no functional change as
a result of this patch.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Mark Brown <broonie@kernel.org>
Cc: Shuah Khan <shuah@kernel.org>
Cc: Will Deacon <will@kernel.org>
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-kselftest@vger.kernel.org