kselftest/arm64: Clarify sve-test preemption comment
The sve-test logic has some dead code with a terse comment.

Remove the dead code and add a clearer comment. The new comment is
indented consistently with other comments. This will make subsequent
removal of other dead code clearer.

There should be no functional change as a result of this patch.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Mark Brown <broonie@kernel.org>
Cc: Shuah Khan <shuah@kernel.org>
Cc: Will Deacon <will@kernel.org>
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-kselftest@vger.kernel.org