kselftest/arm64: Consistently exit fp-stress sub-tests
The fp-stress sub-tests handle failure conditions inconsistently. In
some cases they exit, and in others they send a SIGABRT signal. In some
cases SIGABRT is sent to the entire process group, where it will
terminate the fp-stress harness (potentially before it has dumped any
messages from the failing sub-test). If and when this happens, it's
practically impossible to debug.

For these tests, we gain nothing by terminating the tests using SIGABRT.
It would be simpler and clearer to terminate the tests by exiting with a
non-zero status.

Make the tests exit consistently. Invoke the exit syscall using a macro
to simplify the code and to minimize the risk that the registers are
configured incorrectly for any invocation (as was recently fixed for
several invocations). All existing exit invocations are moved over to
the new macro.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Mark Brown <broonie@kernel.org>
Cc: Shuah Khan <shuah@kernel.org>
Cc: Will Deacon <will@kernel.org>
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-kselftest@vger.kernel.org