kselftest/arm64: Do not enable GCS for the FP stress tests
We enabled GCS for the fp-stress sub-tests in commit:

  bb9ae1a66c85 ("kselftest/arm64: Enable GCS for the FP stress tests")

As noted in that commit message, these tests aren't intended to test
GCS, so this wasn't necessary. While enabling GCS does "give us some
coverage" of GCS, it also makes it harder to attribute any test failure
to a specific cause, and it's more of a hindrance than a help.

If we want better coverage of how GCS interacts with signal handling,
that should live elsewhere, where failures due to GCS are clearly
attributable to GCS.

Stop enabling GCS for the fp-stress sub-tests.

This reverts commit bb9ae1a66c85eeb626864efd812c62026e126ec0.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Mark Brown <broonie@kernel.org>
Cc: Shuah Khan <shuah@kernel.org>
Cc: Will Deacon <will@kernel.org>
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-kselftest@vger.kernel.org